mm/ptshare: Add flag MAP_SHARED_PT to mmap()
When a process mmaps a file with MAP_SHARED flag, it is possible
that any other processes mmaping the same file with MAP_SHARED
flag with same permissions could share the page table entries as
well instead of creating duplicate entries. This patch introduces a
new flag MAP_SHARED_PT for mmap() which a process can use to hint
that it can share page tables with other processes using the same
mapping.

Signed-off-by: Khalid Aziz <khalid.aziz@oracle.com>
[ Allow MAP_PRIVATE to be used with MAP_SHARED_PT as well - Kalesh Singh ]
Signed-off-by: Kalesh Singh <kaleshsingh96@gmail.com>