mm/ptshare: Create new mm struct for page table sharing
When a process passes MAP_SHARED_PT flag to mmap(), create a new mm
struct to hold the shareable page table entries for the newly mapped
region.  This new mm is not associated with a task.  Its lifetime is
until the last shared mapping is deleted.  This patch also adds a
new pointer "ptshare_data" to struct address_space which points to
the data structure that will contain pointer to this newly created
mm along with properties of the shared mapping. ptshare_data
maintains a refcount for the shared mapping so that it can be
cleaned up upon last unmap.

Signed-off-by: Khalid Aziz <khalid.aziz@oracle.com>
Suggested-by: Matthew Wilcox (Oracle) <willy@infradead.org>
[ Allow usage by MAP_PRIVATE,
  Don't allow splitting,
  Rework ptshare vm_flags,
   - Kalesh Singh ]
Signed-off-by: Kalesh Singh <kaleshsingh96@gmail.com>